+++ src/db/migrations/migrations.js
@@ -6,6 +6,7 @@ import m0001 from './0001_items.sql';
 import m0002 from './0002_shops.sql';
 import m0003 from './0003_item_note_urgent.sql';
 import m0004 from './0004_item_not_found.sql';
+import m0005 from './0005_list_look.sql';
 
   export default {
     journal,
@@ -14,7 +15,8 @@ import m0004 from './0004_item_not_found.sql';
 m0001,
 m0002,
 m0003,
-m0004
+m0004,
+m0005
     }
   }